+++ src/commands/CommandHandlerCommon.ts
@@ -9,15 +9,15 @@ import { runInsertCommand } from './insert';
 import { joinVoiceChannelSafe } from './join';
 import { getJoke } from './joke';
 import { runKeysCommand } from './keys';
-import { runMoveItemCommand, moveKeysWrapper } from './move';
+import { moveKeysWrapper, runMoveItemCommand } from './move';
 import { runWhatsPCommand } from './now-playing';
 import { pauseCommandUtil, playCommandUtil, stopPlayingUtil } from './stream/utils';
 import { runPurgeCommand } from './purge';
-import { runRemoveCommand, removePlaylist } from './remove';
+import { removePlaylist, runRemoveCommand } from './remove';
 import { runUniversalSearchCommand } from './search';
 import { runRestartCommand } from './restart';
 import { renameKey, renamePlaylist } from './rename';
-import { runPlayLinkCommand, playLinkNow } from './playLink';
+import { playLinkNow, runPlayLinkCommand } from './playLink';
 import { ZWSP } from '../utils/lib/constants';
 import { runSeekCommand } from './seek';
 import { runRandomToQueue, shuffleQueue } from './playRandomKeys';
