+++ src/commands/join.ts
@@ -1,6 +1,7 @@
 import { Message } from 'discord.js';
 import LocalServer from '../utils/lib/LocalServer';
 import { LEAVE_VC_TIMEOUT } from '../utils/lib/constants';
+
 const { sessionEndEmbed } = require('../utils/embed');
 const { resetSession, botInVC } = require('../utils/utils');
 const processStats = require('../utils/lib/ProcessStats');
