+++ src/commands/lyrics.ts
@@ -3,6 +3,7 @@ import { Message, MessageReaction, TextChannel, User } from 'discord.js';
 import { botID, StreamType } from '../utils/lib/constants';
 import EmbedBuilderLocal from '../utils/lib/EmbedBuilderLocal';
 import ytdl from 'ytdl-core-discord';
+
 const fetch = require('isomorphic-unfetch');
 const { getData } = require('spotify-url-info')(fetch);
 // imports for YouTube captions
