+++ src/commands/playRandomKeys.ts
@@ -1,7 +1,7 @@
 import { Message, TextChannel } from 'discord.js';
 import LocalServer from '../utils/lib/LocalServer';
 import { getAssumptionMultipleMethods } from './search';
-import { verifyPlaylist, createQueueItem, getLinkType, botInVC, setSeamless } from '../utils/utils';
+import { botInVC, createQueueItem, getLinkType, setSeamless, verifyPlaylist } from '../utils/utils';
 import { MAX_QUEUE_S } from '../utils/lib/constants';
 import { getPlaylistItems } from '../utils/playlist';
 import { updateActiveEmbed } from '../utils/embed';
