+++ src/commands/stream/recommendations.ts
@@ -1,7 +1,7 @@
-import { botInVC, setSeamless, resetSession, pushQueue, verifyUrl } from '../../utils/utils';
-import { bot, SPOTIFY_BASE_LINK, CORE_ADM } from '../../utils/lib/constants';
+import { botInVC, pushQueue, resetSession, setSeamless, verifyUrl } from '../../utils/utils';
+import { bot, CORE_ADM, SPOTIFY_BASE_LINK } from '../../utils/lib/constants';
 import { playLinkToVC } from './stream';
-import { updateActiveEmbed, createEmbed } from '../../utils/embed';
+import { createEmbed, updateActiveEmbed } from '../../utils/embed';
 import { addLinkToQueue } from '../../utils/playlist';
 import { isCoreAdmin } from '../../utils/permissions';
 import { Message } from 'discord.js';
