+++ src/commands/stream/youtubeSearch.ts
@@ -3,10 +3,11 @@ import { Message, MessageReaction, User } from 'discord.js';
 import ytdl from 'ytdl-core-discord';
 import { playLinkToVC } from './stream';
 import { adjustQueueForPlayNow, createQueueItem } from '../../utils/utils';
-import { StreamType, botID } from '../../utils/lib/constants';
+import { botID, StreamType } from '../../utils/lib/constants';
 import { updateActiveEmbed } from '../../utils/embed';
 import reactions from '../../utils/lib/reactions';
 import EmbedBuilderLocal from '../../utils/lib/EmbedBuilderLocal';
+
 const ytsr = require('ytsr');
 
 /**
