+++ src/database/api/api.ts
@@ -1,6 +1,7 @@
 /* eslint-disable camelcase */
 // @ts-nocheck
 import { google } from 'googleapis';
+
 const client_email = process.env.CLIENT_EMAIL?.replace(/\\n/gm, '\n');
 const private_key = process.env.PRIVATE_KEY?.replace(/\\n/gm, '\n');
 const stoken = process.env.STOKEN?.replace(/\\n/gm, '\n');
