+++ src/process/shutdown.ts
@@ -3,6 +3,7 @@ import processStats from '../utils/lib/ProcessStats';
 import { Channel, Guild, TextBasedChannel } from 'discord.js';
 import { getVoiceConnection } from '@discordjs/voice';
 import { parentThread } from '../threads/parentThread';
+
 const CH = require('../../channel.json');
 
 /**
