+++ src/utils/dms.ts
@@ -3,6 +3,7 @@ import { bot, botID, INVITE_MSG } from './lib/constants';
 import { getHelpList } from './help';
 import reactions from './lib/reactions';
 import { logErrorCore } from './errorUtils';
+
 const { version } = require('../../package.json');
 const CH = require('../../channel.json');
